requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community>=0.0.28
pypdf==4.1.0
chromadb==0.4.24
sentence-transformers==2.5.1
python-dotenv==1.0.1
pydantic==2.6.3
torch==2.2.1
transformers==4.38.2
datasets==2.18.0
accelerate==0.27.2
bitsandbytes==0.42.0
peft==0.8.2
python-docx==1.1.0 